requirements.txt
ebaysdk
requests
python-dotenv 
beautifulsoup4